refactor(collections): remove deprecated APIs for 9.0.0
Removes the breaking changes for 9.0 from the cdk/collections entry point. Also renames the file with the SelectionModel so it's consistent with the class that it contains.

BREAKING CHANGES:

* `SelectionModel.onChange` has been removed. Use `SelectionModel.changed` instead.
