+++ src/cdk/collections/public-api.ts
@@ -9,7 +9,7 @@
 export * from './array-data-source';
 export * from './collection-viewer';
 export * from './data-source';
-export * from './selection';
+export * from './selection-model';
 export {
   UniqueSelectionDispatcher,
   UniqueSelectionDispatcherListener,
