+++ src/cdk/collections/tree-adapter.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {SelectionModel} from './selection';
+import {SelectionModel} from './selection-model';
 
 
 /**
