+++ src/material/tree/data-source/flat-data-source.ts
@@ -149,7 +149,7 @@ export class MatTreeFlatDataSource<T, F> extends DataSource<F> {
   connect(collectionViewer: CollectionViewer): Observable<F[]> {
     const changes = [
       collectionViewer.viewChange,
-      this._treeControl.expansionModel.onChange,
+      this._treeControl.expansionModel.changed,
       this._flattenedData
     ];
     return merge(...changes).pipe(map(() => {
